app/vminsert/netstorage: Add replication groups to netstorage. Addresses #541.
